Refactor partner registration process and entities
Removed the `FormParceiroCompletarCadastroDto` and `FormUsuarioParceiro` classes. Introduced a new `Parceiro` class with updated properties. Updated `ConnectMataoContext` to reflect changes in the database schema, including the removal of the `FormsUsuarioParceiro` table and the addition of the `TB_Parceiro` table. Created new DTOs and validators for the `Parceiro` entity. Updated endpoints in `Program.cs` to streamline the partner registration process.